Remove tests (for now)
Personally, I dislike the setup of the existing test suite because of how it requires you to make fields and methods mutable or public where they otherwise should not be. Admittedly, I'm not very familiar with Android unit tests, but in the future, I hope to remedy this issue and cover the same tests with another approach.